get next line mas

+++ prueba/libft.h
@@ -6,7 +6,7 @@
 /*   By: irifarac <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/18 12:30:04 by irifarac          #+#    #+#             */
-/*   Updated: 2022/02/12 10:04:55 by irifarac         ###   ########.fr       */
+/*   Updated: 2022/02/16 10:13:00 by irifarac         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -39,7 +39,7 @@ void	*ft_memset(void *dest, int ch, size_t count);
 char	*ft_strchr(const char *str, int ch);
 size_t	ft_strlcat(char *dest, const char *src, size_t size);
 size_t	ft_strlcpy(char *dest, const char *src, size_t size);
-int		ft_strlen(const char *str);
+size_t	ft_strlen(const char *str);
 int		ft_strncmp(const char *str1, const char *str2, size_t number);
 char	*ft_strnstr(const char *dest, const char *src, size_t count);
 char	*ft_strrchr(const char *str, int ch);
